requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.22.0
pandas>=2.0.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0

# Deep learning and model-related
timm>=0.9.2  # For model architectures
optuna>=3.1.0  # For hyperparameter optimization
plotly>=5.14.0  # For visualization in Optuna

# Image processing
Pillow>=9.5.0  # PIL fork for image processing
openslide-python>=1.2.0  # For reading whole slide images
opencv-python>=4.7.0  # For image processing

# Utilities
shapely>=2.0.0  # For geometric operations
tensorboard>=2.12.0  # For logging during training
jupyter>=1.0.0  # For notebooks
ipywidgets>=8.0.0  # For interactive widgets in notebooks

# Testing and development
pytest>=7.3.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0